issue-253 No xcresult output_type for scan
This strips out the xcresult type (and associated filename) from the
options that are being sent to scan when building.

Also improve the documentation around the `xcresult` output_type option
to explain a little more what it is doing and why multi_scan is _not_
sending down the `result_bundle`, and why a symbolic link is being
created if the `result_bundle` option was chosen.