issue-254 Improve batch and parallel testrun logic (#257)
This improves the logic around the creation of simulator clones to
account for a reduced number of tests. If there are fewer tests than the
requested parallel_test_runs, reduce the number of simulators to the
number of tests.

Additionally, if there was only 1 batch requested, stop prefixing
results with the batch number and output directly to the requested
output folder.