Specifically only retry failing tests
In the case of Swift tests with Quick, the initial lists of tests in `only_testing` will include just the `spec` method (i.e `TestClass/spec`), while the junit list of passing / failing tests will include the explicit test case names that were determined during runtime (since the tests cases, for Quick tests, are all derived at runtime and not compile time). This means that the initial `only_testing` options and the `passing` test list will never overlap, and so in the case of any test failure, all tests will be re-run instead of only the failing ones.

A concrete example of the mismatch between `only_testing` and `passing` in my local repo:

> Fastlane::Actions::TestsFromJunitAction.run(config).fetch(:passing, Hash.new).map(&:shellsafe_testidentifier)
=> ["AppUITests/ForgotPasswordUITests/returns_back_to_the_main_screen_after_a_succesful_reset",
 "AppUITests/ForgotPasswordUITests/shows_an_error_when_the_reset_fails",
 "AppUITests/RequestAccountTests/shows_the_request_account_screen_when_Request_Account_link_pressed_and_is_dismissed_when_Done_is_pressed",
 "AppUITests/SessionUITests/prompts_user_for_their_password_if_their_sesison_expired_while_the_app_was_in_the_background",
 "AppUITests/SessionUITests/throws_an_error_if_the_password_is_incorrect"]

> @options[:only_testing]
=> ["AppUITests/ForgotPasswordUITests/spec", "AppUITests/SessionUITests/spec", "AppUITests/OutdatedAppUITest/spec", "AppUITests/RequestAccountTests/spec"]

As you can tell, if we attempt to subtract the `passing` tests according to junit from the `only_testing` flags originally sent through, we will subtract nothing from `only_testing`, causing all tests to be re-run.