issue-305 Add option to override scan_options
This option was added in order to allow users to change the options that
will be passed to the scan fastlane action. This is useful for example,
when users want to remove `xctestrun` option when running tests under
older versions of Xcode with test artifacts built with newer versions of
Xcode.